Conflicts:
	EDDiscovery/Controls/ComboBoxCustom.cs
	EDDiscovery/DB/DistanceClass.cs
	EDDiscovery/DB/SQLiteDBClass.cs
	EDDiscovery/DB/SavedRouteClass.cs
	EDDiscovery/DB/SystemClass.cs
	EDDiscovery/DB/SystemNoteClass.cs
	EDDiscovery/DB/TravelLogUnit.cs
	EDDiscovery/DB/VisitedSystemsClass.cs
	EDDiscovery/DB/WantedSystemClass.cs
	EDDiscovery/EDDB/EDDBClass.cs
	EDDiscovery/EDDiscoveryForm.Designer.cs
	EDDiscovery/EDDiscoveryForm.cs

+++ EDDiscovery/DB/IVisitedSystems.cs
@@ -4,11 +4,11 @@ namespace EDDiscovery2.DB
 {
     public interface IVisitedSystems
     {
-        int id { get; set; }
+        long id { get; set; }
         string Name { get; set; }
         DateTime Time { get; set; }
         int Commander { get; set; }
-        int Source { get; set; }
+        long Source { get; set; }
         string Unit { get; set; }
         bool EDSM_sync { get; set; }
         int MapColour { get; set; }
